fixed shrug channels

+++ MARKBOT/MarkBot-v2/bot.py
@@ -1,5 +1,6 @@
 from discord.ext import commands
 import discord
+from pathlib import Path
 
 l = '''
  ██████╗ ██████╗  ██████╗      ██╗███████╗ ██████╗████████╗    
@@ -21,13 +22,14 @@ print(l)
 with open('config.json') as file:
     config = eval(file.read())
 
-with open('database', 'w+') as file:
-    file.write('{}')
+if not Path('database').is_file():
+    with open('database', 'w+') as file:
+        file.write('{}')
 
 initial_extensions = ['cogs.Functionality']
 
 bot = commands.Bot(
-    command_prefix=config['prefix'], description='An easy to use multipurpose Discord bot!')
+    command_prefix=config['prefix'], description='An easy to use multipurpose Discord bot!', intents=discord.Intents.all())
 
 if __name__ == '__main__':
     for extension in initial_extensions:
